refactoring: move average distribution chart from dashboard to component
